Potential bug fix for joining.
When 2 players simultaneously attempt to join a game with already 3
confirmed players in it, the confirmed players should reject at least
one of them. Therefore, they need to check whether the number of players
which may end up in the game (as opposed to just those that are
currently in the game) doesn't exceed the maximum number of players.